+++ e2e/mock-api-v2/src/endpoints/open-id-configuration.endpoint.ts
@@ -1,7 +1,7 @@
 import { Schema } from '@effect/schema';
 import { pipe } from 'effect';
 import { Api } from 'effect-http';
-import { openIdConfigurationResponseSchema } from '../schemas/openid/openidConfigurationResponse';
+import { openIdConfigurationResponseSchema } from '../schemas/open-id-configuration/open-id-configuration-response.schema';
 
 const openidConfiguration = Api.addEndpoint(
   pipe(
