+++ e2e/mock-api-v2/src/handlers/open-id-configuration.handler.ts
@@ -1,7 +1,7 @@
 import { Effect } from 'effect';
 import { RouterBuilder } from 'effect-http';
 import { apiSpec } from '../spec';
-import { openidConfigurationResponse } from '../responses/openidConfiguration';
+import { openidConfigurationResponse } from '../responses/open-id-configuration';
 
 const openidConfiguration = RouterBuilder.handler(apiSpec, 'openidConfiguration', () =>
   // eslint-disable-next-line require-yield
