+++ e2e/mock-api-v2/src/handlers/token.handler.ts
@@ -1,7 +1,7 @@
 import { Effect } from 'effect';
 import { RouterBuilder } from 'effect-http';
 import { apiSpec } from '../spec';
-import { Tokens } from '../services/tokens';
+import { Tokens } from '../services/tokens.service';
 
 const tokenHandler = RouterBuilder.handler(apiSpec, 'PingOneToken', () =>
   Effect.gen(function* () {
