+++ e2e/mock-api-v2/src/handlers/userinfo.handler.ts
@@ -1,7 +1,7 @@
 import { Effect } from 'effect';
 import { RouterBuilder } from 'effect-http';
 import { apiSpec } from '../spec';
-import { UserInfo } from '../services/userinfo';
+import { UserInfo } from '../services/userinfo.service';
 
 const userInfoHandler = RouterBuilder.handler(apiSpec, 'UserInfo', (request, security) =>
   Effect.gen(function* () {
