--- e2e/mock-api-v2/src/helpers/routeTypes.ts
@@ -1,8 +1,0 @@
-type Routes =
-  | '/userinfo'
-  | '/tokens'
-  | '/authorize'
-  | '/openidConfiguration'
-  | '/customHtmlTemplate';
-
-export { type Routes };
