+++ e2e/mock-api-v2/src/responses/custom-html-template/ping-protect-node.ts
