+++ e2e/mock-api-v2/src/responses/open-id-configuration.ts
