+++ e2e/mock-api-v2/src/responses/return-success-redirect.ts
