+++ e2e/mock-api-v2/src/schemas/custom-html-template/custom-html-template-request.schema.ts
