+++ e2e/mock-api-v2/src/schemas/open-id-configuration/open-id-configuration-response.schema.ts
