+++ e2e/mock-api-v2/src/schemas/return-success-response-redirect.schema.ts
