+++ e2e/mock-api-v2/src/schemas/token/token.schema.ts
