+++ e2e/mock-api-v2/src/schemas/userinfo/userinfo.schema.ts
